Creation of Hgmd files

Create a positive variant file from the Hgmd data.

Usage

Hgmd_divide(hgmd_file_name)

Arguments

hgmd_file_name

Path to hgmd file name

Details

Based on HGMD data, list amino acid changes (column: HGVSprotein), VCF (column: VCF), DM if DM in HGMD, DMp if DM? (column: possible), Pathogenic if pathogenic but not a target phenotype, and Target if target phenotype(column: Type) for each variant and save it. This is processed by Hgmd_divide and can be processed by Edit_variant_data.

Value

References

Author

Yuya Hatano

Note

See also

Examples

Hgmd_divide("./source/TARDBP.csv")
Edit_gnomAD_file("./source/gnomAD_v3.1.2_ENST00000240185_2023_02_14_13_23_48.csv", "./source/TARDBP_gnomAD.csv")
Edit_variant_data("../CADD_REVEL/AlphScore_final.tsv", "./source/TARDBP.csv", "./source/TARDBP_gnomAD.csv", "Q13148", "TARDBP", "./source/TARDBPvariantdata.csv")
Edit_polyphen_data("../dbNSFP/dbNSFP4.3a/dbNSFP4.3a_variant.chr1","./source/TARDBPvariantdata.csv", "TARDBP_alph.csv",  "Q13148",  11012654, 11025492 ,"./source/TARDBPvariantdatapol.csv", "./source/TARDBP_alphpol.csv")
Edit_final_variant_file("./source/TARDBP_alphpol.csv","./source/TARDBP_alphpol2.csv")
MOVA("./source/Q13148.fa", "TARDBP", "./source/AF-Q13148-F1-model_v2.pdb","./source/TARDBP_alphpol2.csv", "./source/TARDBPvariantdatapol.csv")

## The function is currently defined as
function (hgmd_file_name) 
{
    x <- fread(hgmd_file_name)
    x <- x %>% separate(VCF, c("buf1", "buf2"), sep = ":")
    x <- x %>% separate(buf2, c("buf3", "Alternate"), sep = "/")
    x$Chromosome <- str_sub(x$buf1, 4, -1)
    x$Position <- str_sub(x$buf3, 1, -2)
    x$Reference <- str_sub(x$buf3, -1, -1)
    x[, `:=`(buf1, NULL)]
    x[, `:=`(buf3, NULL)]
    fwrite(x, hgmd_file_name)
  }
#> function (hgmd_file_name) 
#> {
#>     x <- fread(hgmd_file_name)
#>     x <- x %>% separate(VCF, c("buf1", "buf2"), sep = ":")
#>     x <- x %>% separate(buf2, c("buf3", "Alternate"), sep = "/")
#>     x$Chromosome <- str_sub(x$buf1, 4, -1)
#>     x$Position <- str_sub(x$buf3, 1, -2)
#>     x$Reference <- str_sub(x$buf3, -1, -1)
#>     x[, `:=`(buf1, NULL)]
#>     x[, `:=`(buf3, NULL)]
#>     fwrite(x, hgmd_file_name)
#>   }
#> <environment: 0x0000011a816a0900>